% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStateSpace-methods.R
\name{coef.bootstatespace}
\alias{coef.bootstatespace}
\title{Estimated Parameter Method for an Object of Class
\code{bootstatespace}}
\usage{
\method{coef}{bootstatespace}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{bootstatespace}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of estimated parameters.
}
\description{
Estimated Parameter Method for an Object of Class
\code{bootstatespace}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
