% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStateSpace-methods.R
\name{extract.bootstatespace}
\alias{extract.bootstatespace}
\title{Extract Method for an Object of Class
\code{bootstatespace}}
\usage{
\method{extract}{bootstatespace}(object, what = NULL)
}
\arguments{
\item{object}{Object of Class \code{bootstatespace}.}

\item{what}{Character string.
What specific matrix to extract.
If \code{what = NULL},
extract all available matrices.}
}
\value{
Returns a list.
Each element of the list
is a list of bootstrap estimates
in matrix format.
}
\description{
Extract Method for an Object of Class
\code{bootstatespace}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
