% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStateSpace-methods.R
\name{print.bootstatespace}
\alias{print.bootstatespace}
\title{Print Method for an Object of Class
\code{bootstatespace}}
\usage{
\method{print}{bootstatespace}(x, alpha = NULL, type = "pc", digits = 4, ...)
}
\arguments{
\item{x}{Object of Class \code{bootstatespace}.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.
If \code{alpha = NULL},
use the argument \code{alpha} used in \code{x}.}

\item{type}{Charater string.
Confidence interval type, that is,
\code{type = "pc"} for percentile;
\code{type = "bc"} for bias corrected.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Prints a matrix of
estimates,
standard errors,
number of bootstrap replications,
and
confidence intervals.
}
\description{
Print Method for an Object of Class
\code{bootstatespace}
}
\examples{
\dontrun{
# prepare parameters
## number of individuals
n <- 5
## time points
time <- 50
## dynamic structure
p <- 3
mu0 <- rep(x = 0, times = p)
sigma0 <- 0.001 * diag(p)
sigma0_l <- t(chol(sigma0))
alpha <- rep(x = 0, times = p)
beta <- 0.50 * diag(p)
psi <- 0.001 * diag(p)
psi_l <- t(chol(psi))

path <- tempdir()

pb <- PBSSMVARFixed(
  R = 10L, # use at least 1000 in actual research
  path = path,
  prefix = "var",
  n = n,
  time = time,
  mu0 = mu0,
  sigma0_l = sigma0_l,
  alpha = alpha,
  beta = beta,
  psi_l = psi_l,
  type = 0,
  ncores = 1, # consider using multiple cores
  seed = 42
)
print(pb)
print(pb, type = "bc")
}

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
