% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStateSpace-methods.R
\name{summary.bootstatespace}
\alias{summary.bootstatespace}
\title{Summary Method for an Object of Class
\code{bootstatespace}}
\usage{
\method{summary}{bootstatespace}(object, alpha = NULL, type = "pc", digits = 4, ...)
}
\arguments{
\item{object}{Object of Class \code{bootstatespace}.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.
If \code{alpha = NULL},
use the argument \code{alpha} used in \code{object}.}

\item{type}{Charater string.
Confidence interval type, that is,
\code{type = "pc"} for percentile;
\code{type = "bc"} for bias corrected.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of
estimates,
standard errors,
number of bootstrap replications,
and
confidence intervals.
}
\description{
Summary Method for an Object of Class
\code{bootstatespace}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
