% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStateSpace-methods.R
\name{vcov.bootstatespace}
\alias{vcov.bootstatespace}
\title{Sampling Variance-Covariance Matrix Method for an Object of Class
\code{bootstatespace}}
\usage{
\method{vcov}{bootstatespace}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{bootstatespace}.}

\item{...}{additional arguments.}
}
\value{
Returns the variance-covariance matrix of estimates.
}
\description{
Sampling Variance-Covariance Matrix Method for an Object of Class
\code{bootstatespace}
}
\examples{
\dontrun{
# prepare parameters
## number of individuals
n <- 5
## time points
time <- 50
## dynamic structure
p <- 3
mu0 <- rep(x = 0, times = p)
sigma0 <- 0.001 * diag(p)
sigma0_l <- t(chol(sigma0))
alpha <- rep(x = 0, times = p)
beta <- 0.50 * diag(p)
psi <- 0.001 * diag(p)
psi_l <- t(chol(psi))

path <- tempdir()

pb <- PBSSMVARFixed(
  R = 10L, # use at least 1000 in actual research
  path = path,
  prefix = "var",
  n = n,
  time = time,
  mu0 = mu0,
  sigma0_l = sigma0_l,
  alpha = alpha,
  beta = beta,
  psi_l = psi_l,
  type = 0,
  ncores = 1, # consider using multiple cores
  seed = 42
)
vcov(pb)
}

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
