% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStateSpace-methods.R
\name{confint.bootstatespace}
\alias{confint.bootstatespace}
\title{Confidence Intervals Method for an Object of Class
\code{bootstatespace}}
\usage{
\method{confint}{bootstatespace}(object, parm = NULL, level = 0.95, type = "pc", ...)
}
\arguments{
\item{object}{Object of Class \code{bootstatespace}.}

\item{parm}{a specification of which parameters
are to be given confidence intervals,
either a vector of numbers or a vector of names.
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{type}{Charater string.
Confidence interval type, that is,
\code{type = "pc"} for percentile;
\code{type = "bc"} for bias corrected.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of confidence intervals.
}
\description{
Confidence Intervals Method for an Object of Class
\code{bootstatespace}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
