% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStateSpace-methods.R
\name{extract.bootstatespace}
\alias{extract.bootstatespace}
\title{Extract Method for an Object of Class
\code{bootstatespace}}
\usage{
\method{extract}{bootstatespace}(object, what = NULL)
}
\arguments{
\item{object}{Object of Class \code{bootstatespace}.}

\item{what}{Character string.
What specific matrix to extract.
If \code{what = NULL},
extract all available matrices.}
}
\value{
Returns a list.
Each element of the list
is a list of bootstrap estimates
in matrix format.
}
\description{
Extract Method for an Object of Class
\code{bootstatespace}
}
\examples{
\dontrun{
# prepare parameters
## number of individuals
n <- 5
## time points
time <- 50
## dynamic structure
p <- 3
mu0 <- rep(x = 0, times = p)
sigma0 <- 0.001 * diag(p)
sigma0_l <- t(chol(sigma0))
alpha <- rep(x = 0, times = p)
beta <- 0.50 * diag(p)
psi <- 0.001 * diag(p)
psi_l <- t(chol(psi))

path <- tempdir()

pb <- PBSSMVARFixed(
  R = 10L, # use at least 1000 in actual research
  path = path,
  prefix = "var",
  n = n,
  time = time,
  mu0 = mu0,
  sigma0_l = sigma0_l,
  alpha = alpha,
  beta = beta,
  psi_l = psi_l,
  type = 0,
  ncores = 1, # consider using multiple cores
  seed = 42
)
extract(pb, what = "beta")
}

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
