% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStateSpace-methods.R
\name{vcov.bootstatespace}
\alias{vcov.bootstatespace}
\title{Sampling Variance-Covariance Matrix Method for an Object of Class
\code{bootstatespace}}
\usage{
\method{vcov}{bootstatespace}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{bootstatespace}.}

\item{...}{additional arguments.}
}
\value{
Returns the variance-covariance matrix of estimates.
}
\description{
Sampling Variance-Covariance Matrix Method for an Object of Class
\code{bootstatespace}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
